// src/modules/register/interfaces/Clientes.interface.ts
export interface Clientes {
    IdCliente: number;
    Nombre: string;
    RFC?: string;
    Sucursal?: string;
    Direccion?: string;
    Colonia?: string;
    Poblacion?: string;
    Estado?: string;
    CP?: string;
    Telefono?: string;
    Telefono2?: string;
    Telefono3?: string;
    eMail?: string;
    Credito?: number;
    Contacto?: string;
    ManoObra?: number;
    Descuento?: number;
}